+++ clients/main/package.json
@@ -20,6 +20,7 @@
     "deep-equal": "^2.2.3",
     "lodash": "^4.17.21",
     "notistack": "^3.0.1",
+    "pithekos-lib": "^0.0.8",
     "prop-types": "^15.8.1",
     "react": "^17.0.0 || ^18.0.0",
     "react-dom": "^17.0.0 || ^18.0.0",
