--- clients/main/src/contexts/bcv.js
@@ -1,5 +1,0 @@
-import {createContext} from 'react';
-
-const BcvContext = createContext({systemBcv: false, setSystemBcv: ()=>{}});
-
-export default BcvContext;
