--- clients/main/src/contexts/debug.js
@@ -1,5 +1,0 @@
-import {createContext} from 'react';
-
-const DebugContext = createContext({debug: false, setDebug: ()=>{}});
-
-export default DebugContext;
