--- clients/main/src/contexts/i18n.js
@@ -1,5 +1,0 @@
-import {createContext} from 'react';
-
-const I18nContext = createContext({});
-
-export default I18nContext;
