--- clients/main/src/contexts/messages.js
@@ -1,5 +1,0 @@
-import {createContext} from 'react';
-
-const MessagesContext = createContext({messages: [], setMessages: ()=>{}});
-
-export default MessagesContext;
