--- clients/main/src/contexts/net.js
@@ -1,5 +1,0 @@
-import {createContext} from 'react';
-
-const NetContext = createContext({enableNet: false, setEnableNet: ()=>{}});
-
-export default NetContext;
