--- clients/main/src/lib/i18n.js
@@ -1,3 +1,0 @@
-const doI18n = (key, source) => source[key] || key;
-
-export {doI18n};
