+++ clients/main/src/pages/LocalProjects/DeleteProjectButton.jsx
@@ -1,8 +1,7 @@
 import {useContext} from 'react';
 import {IconButton} from "@mui/material";
 import DeleteIcon from '@mui/icons-material/Delete';
-import {getJson} from "../../lib/get";
-import DebugContext from "../../contexts/debug";
+import {debugContext as DebugContext, getJson} from "pithekos-lib";
 
 function DeleteProjectButton({project}) {
     const {debugRef} = useContext(DebugContext);
