+++ clients/main/src/pages/Workspace/Workspace.js
@@ -1,17 +1,13 @@
-// import {useContext} from "react";
+import React from 'react';
 import {useLocation} from "react-router-dom";
-import PtksPage from "../../components/PtksPage";
 import WorkspaceCard from "./WorkspaceCard";
-import BcvPicker from "./BcvPicker";
-// import I18nContext from "../../contexts/i18n";
-// import {doI18n} from "../../lib/i18n";
-
-import React from 'react'
+// import BcvPicker from "./BcvPicker";
 import {
     createTilePanes,
     TileContainer,
     TileProvider,
 } from 'react-tile-pane'
+import {Header} from "pithekos-lib";
 
 
 const paneStyle = {
@@ -21,7 +17,6 @@ const paneStyle = {
 }
 const Workspace = () => {
     const locationState = Object.entries(useLocation().state);
-    // const i18n = useContext(I18nContext);
     const resources = locationState
         .map(kv => {
             return {...kv[1], local_path: kv[0]}
@@ -53,11 +48,11 @@ const Workspace = () => {
         rootPane.children.pop();
     }
     const paneList = createTilePanes(tileElements)[0];
-    return <PtksPage
-        subtitleKey="pages:workspace:subtitle"
-        margin={0}
-        widget={<BcvPicker/>}
-    >
+    return <>
+        <Header
+            subtitleKey="pages:workspace:subtitle"
+            requireNet={false}
+        />
         <TileProvider
             tilePanes={paneList}
             rootNode={rootPane}
@@ -66,6 +61,6 @@ const Workspace = () => {
                 <TileContainer/>
             </div>
         </TileProvider>
-    </PtksPage>
+    </>
 }
 export default Workspace;
